long int

+++ bandit_nash_test.cpp
@@ -188,7 +188,7 @@ void update_ucb(std::vector<std::vector<double>>& infoset_ucb_values, std::vecto
 }
 
 
-void calc_nash_ucb(long int num_iterations, std::vector<std::string>& x_information_sets, std::vector<std::string>& o_information_sets,  int log_flag, int log_frequency, int C, PolicyVec& x_ucb_policy, PolicyVec& o_ucb_policy) {
+void calc_nash_ucb(long int num_iterations, std::vector<std::string>& x_information_sets, std::vector<std::string>& o_information_sets,  int log_flag, long int log_frequency, int C, PolicyVec& x_ucb_policy, PolicyVec& o_ucb_policy) {
     std::vector<long int> x_infoset_time_steps(x_information_sets.size(), 0);
     std::vector<std::vector<double>> x_infoset_ucb_values(x_information_sets.size(), std::vector<double>(13, 0.0));
     std::vector<std::vector<double>> x_infoset_empirical_reward(x_information_sets.size(), std::vector<double>(13, 0.0));
